+++ packages/fans-society-contracts/contracts/tokens/interfaces/IProjectTokenERC20.sol
@@ -4,11 +4,13 @@ pragma solidity 0.8.17;
 import { IERC20MetadataUpgradeable } from '@openzeppelin/contracts-upgradeable/token/ERC20/extensions/IERC20MetadataUpgradeable.sol';
 
 interface IProjectTokenERC20 is IERC20MetadataUpgradeable {
-	
-	event TokenClaimed(address account, uint amount);
+	event TokenClaimed(address account, uint256 amount);
 
-	function safeTransferFrom(address from, address to, uint256 value) external;
+	function safeTransferFrom(
+		address from,
+		address to,
+		uint256 value
+	) external;
 
 	function claim(address account, uint256 amount) external;
-
 }
