+++ packages/fans-society-contracts/contracts/tokens/interfaces/IProjectTokenFactory.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.17;
 
-import {IProjectTokenERC20} from './IProjectTokenERC20.sol';
+import { IProjectTokenERC20 } from './IProjectTokenERC20.sol';
 
 interface IProjectTokenFactory {
 	function createToken(
@@ -9,7 +9,7 @@ interface IProjectTokenFactory {
 		string memory _name,
 		string memory _symbol,
 		address _amm,
-		uint40 _totalSupply,
-		uint40 _initialSupply
+		uint112 _totalSupply,
+		uint112 _initialSupply
 	) external returns (address);
 }
